+++ package.json
@@ -6,18 +6,13 @@
     "@aws-amplify/auth": "^6.3.17",
     "@aws-amplify/backend": "^1.2.0",
     "@aws-amplify/ui-react": "^6.2.0",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
-    "@mui/material": "^6.0.2",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "react": "^18.3.1",
-    "react-cookie": "^7.2.0",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.1",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
@@ -46,7 +41,6 @@
     ]
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "tailwindcss": "^3.4.10"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
   }
 }
